requirements.txt
Django>=4.0
djangorestframework
django-cors-headers
psycopg2-binary
pymongo
neo4j
google-generativeai
PyMuPDF
numpy
boto3
python-dotenv
gunicorn